pyproject.toml
[tool.poetry]
name = "schedule-tgbot"
version = "0.1.0"
description = "Telegram bot for sending actual schedule for sport-school sections."
authors = ["Ilia Kaziamov <kaziamov@outlook.com>"]
readme = "README.md"
packages = [{include = "schedule_tgbot"}]

[tool.poetry.dependencies]
python = "^3.8.1"
aiogram = "^2.25.1"
python-dotenv = "^0.21.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
